Rename network to protocol.
